+++ dask_geomodeling/config.py
@@ -3,6 +3,7 @@ import os
 
 defaults = {
     "root": os.getcwd(),
+    "strict-file-paths": False,
     "raster-limit": 12 * (1024 ** 2),  # ca. 100 MB of float64
     "geometry-limit": 10000,
 }